fix: Prevent config object reference issue in user credential assignment
There was a slim chance that one node would assign its credentials to
the config before the other one would use it for connections, in which
case the user/pass of the other node was assigned to the previous one.
This turned out to be crucial when the auth node would be affected,
since no other node would ever authenticate then.